move header to include/mapbox folder
