+++ include/mapbox/pixelmatch.hpp
